Add files via upload
In this commit I have added all necessary  files for the first page, named index, that will welcome a guest user when opening the webpage. he will be greeted with a small bio about the sites purpose as well as its functionalities explained in compact and sufficient detail. The first pages purpose is to draw the guest client in towards the page, further creating an account or logging in to enjoy what the program has to offer.