Uploaded the discography page
The discography page of Orpheus will contain all albums by a certain artist. these albums can be sorted by a variety of attributes such as length, name, rating and so on. on the right side of the page a picture and a small biography of the artist can be seen as a small insight of who he/she who wrote these albums is.